Code Update

+++ src/SimpleStorage.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.18; // version of solidity (^) greater than 0.8.18
 contract SimpleStorage {
     // Basic Types: boolean, uint, int, address, bytes, string
     bool hasFavourite = false; // default value is false
-    uint myFavouriteNum = 88; // if no value assigned then default is 0;
+    uint256 myFavouriteNum = 88; // if no value assigned then default is 0;
     string favouriteString = "Amey";
     address myAddr = 0x4E5C8a1Af0945983d59f621e1aD6DA7d7D80ff42;
     bytes32 favour = "cat";
@@ -17,6 +17,7 @@ contract SimpleStorage {
         uint256 favouriteNum;
         string name;
     }
+
     Person public myFriend = Person(7, "Omkar");
 
     // Arrays of structure type
